Refactor: Create a DefaultHostFinder
