+++ BrowserLog.NLog/BrowserConsoleTarget.cs
@@ -2,6 +2,7 @@
 using System.Net;
 using System.Net.Sockets;
 using System.Threading;
+using BrowserLog.Commom;
 using BrowserLog.TinyServer;
 using NLog;
 using NLog.Targets;
@@ -42,25 +43,10 @@ namespace BrowserLog.NLog
             {
                 if (Host == null)
                 {
-                    Host = FindLocalIp();
+                    Host = DefaultHostFinder.FindLocalIp();
                 }
                 _channel = _channelFactory.Create(Host, Port, Buffer);
             }
-        }
-
-        // hack described here
-        // http://stackoverflow.com/a/27376368
-        //
-        private static string FindLocalIp()
-        {
-            string localIP;
-            using (Socket socket = new Socket(AddressFamily.InterNetwork, SocketType.Dgram, 0))
-            {
-                socket.Connect("10.0.2.4", 65530);
-                IPEndPoint endPoint = socket.LocalEndPoint as IPEndPoint;
-                localIP = endPoint.Address.ToString();
-            }
-            return localIP;
         }
 
         protected override void Write(LogEventInfo logEvent)
