+++ BrowserLog.log4net/BrowserConsoleAppender.cs
@@ -6,6 +6,7 @@ using System.Net.Sockets;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
+using BrowserLog.Commom;
 using BrowserLog.TinyServer;
 using log4net.Appender;
 using log4net.Core;
@@ -44,7 +45,7 @@ namespace BrowserLog
             {
                 if (Host == null)
                 {
-                    Host = FindLocalIp();
+                    Host = DefaultHostFinder.FindLocalIp();
                 }
                 _channel = _channelFactory.Create(Host, Port, Buffer);
             }
@@ -63,21 +64,6 @@ namespace BrowserLog
             {
                 _channel.Dispose();    
             }
-        }
-
-        // hack described here
-        // http://stackoverflow.com/a/27376368
-        //
-        private static string FindLocalIp()
-        {
-            string localIP;
-            using (Socket socket = new Socket(AddressFamily.InterNetwork, SocketType.Dgram, 0))
-            {
-                socket.Connect("10.0.2.4", 65530);
-                IPEndPoint endPoint = socket.LocalEndPoint as IPEndPoint;
-                localIP = endPoint.Address.ToString();
-            }
-            return localIP;
         }
     }
 }
